requirements.txt
ipykernel
langgraph
langchain
langchain-openai
python-dotenv
dotenv
streamlit
pillow
